Supports the newest version of Aeron. Updated error log usage and notifications. Refactored RMI API (now RMI is accessible via '.rmi' calls). Updated unit tests. Added 'delete' to rmi methods (opposite of 'create').

+++ test/dorkboxTest/network/rmi/multiJVM/TestServer.kt
@@ -54,7 +54,7 @@ object TestServer {
             logger.error("Received finish signal for test for: Client -> Server")
             val `object` = m.testCow
             val id = `object`.id()
-            Assert.assertEquals(124123, id.toLong())
+            Assert.assertEquals(124123, id)
             logger.error("Finished test for: Client -> Server")
 
 //
